UInt -> UInt64

+++ Adamant/Models/Transaction.swift
@@ -9,21 +9,21 @@
 import Foundation
 
 struct Transaction {
-	let id: UInt
+	let id: UInt64
 	let height: Int
-	let blockId: UInt
+	let blockId: UInt64
 	let type: TransactionType
-	let timestamp: UInt
+	let timestamp: UInt64
 	let senderPublicKey: String
 	let senderId: String
 	let requesterPublicKey: String?
 	let recipientId: String
 	let recipientPublicKey: String?
-	let amount: UInt
-	let fee: UInt
+	let amount: UInt64
+	let fee: UInt64
 	let signature: String
 	let signSignature: String?
-	let confirmations: UInt
+	let confirmations: UInt64
 	let signatures: [String]
 	let asset: TransactionAsset
 	
@@ -54,19 +54,19 @@ extension Transaction: Decodable {
 	init(from decoder: Decoder) throws {
 		let container = try decoder.container(keyedBy: CodingKeys.self)
 		
-		self.id = UInt(try container.decode(String.self, forKey: .id))!
+		self.id = UInt64(try container.decode(String.self, forKey: .id))!
 		self.height = try container.decode(Int.self, forKey: .height)
-		self.blockId = UInt(try container.decode(String.self, forKey: .blockId))!
+		self.blockId = UInt64(try container.decode(String.self, forKey: .blockId))!
 		self.type = try container.decode(TransactionType.self, forKey: .type)
-		self.timestamp = try container.decode(UInt.self, forKey: .timestamp)
+		self.timestamp = try container.decode(UInt64.self, forKey: .timestamp)
 		self.senderPublicKey = try container.decode(String.self, forKey: .senderPublicKey)
 		self.senderId = try container.decode(String.self, forKey: .senderId)
 		self.recipientId = try container.decode(String.self, forKey: .recipientId)
 		self.recipientPublicKey = try? container.decode(String.self, forKey: .recipientPublicKey)
-		self.amount = try container.decode(UInt.self, forKey: .amount)
-		self.fee = try container.decode(UInt.self, forKey: .fee)
+		self.amount = try container.decode(UInt64.self, forKey: .amount)
+		self.fee = try container.decode(UInt64.self, forKey: .fee)
 		self.signature = try container.decode(String.self, forKey: .signature)
-		self.confirmations = (try? container.decode(UInt.self, forKey: .confirmations)) ?? 0
+		self.confirmations = (try? container.decode(UInt64.self, forKey: .confirmations)) ?? 0
 		self.requesterPublicKey = try? container.decode(String.self, forKey: .requesterPublicKey)
 		self.signSignature = try? container.decode(String.self, forKey: .signSignature)
 		self.signatures = try container.decode([String].self, forKey: .signatures)
