+++ AdamantTests/FeeCalculatorTests.swift
@@ -11,8 +11,8 @@ import XCTest
 
 class HardFeeCalculatorTests: XCTestCase {
     func testTransferFee() {
-		let amountToTransfer: UInt = 50000000000000000
-		let estimatedFee: UInt = 50000000
+		let amountToTransfer: UInt64 = 50000000000000000
+		let estimatedFee: UInt64 = 50000000
 		
 		let calculatedFee = AdamantFeeCalculator.estimatedFeeFor(transfer: amountToTransfer)
 		XCTAssertEqual(estimatedFee, calculatedFee)
@@ -20,7 +20,7 @@ class HardFeeCalculatorTests: XCTestCase {
 	
 	func testShortMessageFee() {
 		let message = "A quick brown fox bought bitcoins in 2009. Good for you, mr fox. You quick brown mother fucker."
-		let estimatedFee: UInt = 100000
+		let estimatedFee: UInt64 = 100000
 		
 		let calculatedFee = AdamantFeeCalculator.estimatedFeeFor(message: AdamantMessage.text(message))
 		XCTAssertEqual(estimatedFee, calculatedFee)
@@ -33,7 +33,7 @@ The olfactory system is reduced, suggesting that the sperm whale has a poor sens
 By contrast, the auditory system is enlarged.
 The pyramidal tract is poorly developed, reflecting the reduction of its limbs.
 """
-		let estimatedFee: UInt = 200000
+		let estimatedFee: UInt64 = 200000
 		
 		let calculatedFee = AdamantFeeCalculator.estimatedFeeFor(message: AdamantMessage.text(message))
 		XCTAssertEqual(estimatedFee, calculatedFee)
@@ -88,7 +88,7 @@ I'll tell you when
 Mastodon / The Hunter / All The Heavy Lifting
 Brann Timothy Dailor / Troy Jayson Sanders / William Breen Kelliher / William Brent Hinds
 """
-		let estimatedFee: UInt = 400000
+		let estimatedFee: UInt64 = 400000
 		
 		let calculatedFee = AdamantFeeCalculator.estimatedFeeFor(message: AdamantMessage.text(message))
 		XCTAssertEqual(estimatedFee, calculatedFee)
